Refactor TeamZipExport code (move it into special file)
